requirements.txt
asgiref==3.6.0
Django==3.2.18
django-materializecss-form==1.1.17
django-multiselectfield==0.1.12
django-tinymce==3.6.1
numpy==1.21.6
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
typing-extensions==4.5.0
